cloud code/requirements.txt
Flask
requests
google-cloud-storage
google-cloud-pubsub
gunicorn
